+++ src/api/invidious.rs
@@ -8,6 +8,7 @@ use async_trait::async_trait;
 use rand::prelude::*;
 use reqwest::Client;
 use serde_json::Value;
+use std::collections::HashSet;
 use std::sync::Arc;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::time::Duration;
@@ -16,11 +17,7 @@ const API_BACKEND: ApiBackend = ApiBackend::Invidious;
 
 impl From<Value> for ChannelFeed {
     fn from(mut value: Value) -> Self {
-        let mut channel_feed = Self {
-            channel_title: None,
-            channel_id: None,
-            videos: Vec::new(),
-        };
+        let mut channel_feed = Self::default();
 
         let videos = if value["videos"].is_null() {
             value
@@ -30,7 +27,6 @@ impl From<Value> for ChannelFeed {
 
         if let Some(video) = videos.get(0) {
             channel_feed.channel_title = Some(video["author"].as_str().unwrap().to_string());
-
             channel_feed.videos = Video::vec_from_json(videos);
         }
 
@@ -42,6 +38,7 @@ impl From<Value> for ChannelFeed {
 pub struct Instance {
     pub domain: String,
     client: Client,
+    continuation: Option<String>,
     old_version: Arc<AtomicBool>,
 }
 
@@ -58,6 +55,7 @@ impl Instance {
         Self {
             domain,
             client,
+            continuation: None,
             old_version: Arc::new(AtomicBool::new(false)),
         }
     }
@@ -106,6 +104,31 @@ impl Instance {
 
         Ok(Video::vec_from_json(videos_array))
     }
+
+    async fn get_more_videos_helper(&mut self, channel_id: &str) -> Result<Vec<Video>> {
+        let url = format!("{}/api/v1/channels/{}/videos", self.domain, channel_id,);
+        let mut query = vec![(
+            "fields",
+            "videos(title,videoId,published,publishedText,lengthSeconds,isUpcoming,premiereTimestamp)",
+        )];
+
+        let continuation_token;
+
+        if let Some(token) = &self.continuation {
+            continuation_token = token.to_owned();
+            query.push(("continuation", &continuation_token));
+        }
+
+        let response = self.client.get(&url).query(&query).send().await?;
+        let mut value = response.error_for_status()?.json::<Value>().await?;
+
+        self.continuation = value
+            .get("continuation")
+            .and_then(Value::as_str)
+            .map(ToString::to_string);
+
+        Ok(Video::vec_from_json(value["videos"].take()))
+    }
 }
 
 #[async_trait]
@@ -125,11 +148,7 @@ impl Api for Instance {
     }
 
     async fn get_videos_of_channel(&mut self, channel_id: &str) -> Result<ChannelFeed> {
-        let mut channel_feed = ChannelFeed {
-            channel_title: None,
-            channel_id: Some(channel_id.to_string()),
-            videos: Vec::new(),
-        };
+        let mut channel_feed = ChannelFeed::new(channel_id);
 
         if OPTIONS.videos_tab
             && let Ok(videos) = self
@@ -235,6 +254,41 @@ impl Api for Instance {
         let response = self.client.get(&url).send().await?.error_for_status()?;
 
         Ok(quick_xml::de::from_str(&response.text().await?)?)
+    }
+
+    async fn get_more_videos(
+        &mut self,
+        channel_id: &str,
+        present_videos: HashSet<String>,
+    ) -> Result<ChannelFeed> {
+        let mut feed = ChannelFeed {
+            channel_title: None,
+            channel_id: Some(channel_id.to_owned()),
+            videos: self.get_more_videos_helper(channel_id).await?,
+        };
+
+        let new_video_present = |videos: &[Video]| {
+            !videos
+                .iter()
+                .all(|video| present_videos.contains(&video.video_id))
+        };
+
+        if new_video_present(&feed.videos) {
+            return Ok(feed);
+        }
+
+        while self.continuation.is_some()
+            && let Ok(videos) = self.get_more_videos_helper(channel_id).await
+        {
+            let new = new_video_present(&videos);
+            feed.extend_videos(videos);
+
+            if new {
+                return Ok(feed);
+            }
+        }
+
+        Ok(ChannelFeed::default())
     }
 
     async fn get_video_formats(&self, video_id: &str) -> Result<VideoInfo> {